jQuery
A button that changes several CSS properties of an element on the page at the same time
A button that inserts a message in a box at the top of your page - an custom alert
An animation with a custom timing
A form with two input fields. If you click on a field and then click outside of it before you type anything there - an alert should appear under that field.